Aula 12/08/2016

+++ HO06/src/model/Cliente.java
@@ -4,7 +4,17 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 
 @Entity
 @Table(name="CLIENTES")
@@ -12,7 +22,7 @@ public class Cliente {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "CLIENTE_ID")
-	private long custId;
+	private int	 custId;
 
 	@ManyToMany(mappedBy = "clientes")
 	private Collection<Reserva> reservas = new ArrayList<Reserva>();
